+++ packages/ui-kit/src/components/message-input/index.tsx
@@ -1,5 +1,6 @@
-import { ComponentProps, ComponentPropsWithoutRef, forwardRef } from 'react'
+import { ComponentPropsWithoutRef, ElementRef, forwardRef } from 'react'
 
+import { TextareaAutosize, TextareaAutosizeProps } from '../textarea-autosize'
 import { SendMessageIcon } from './send-message-icon'
 import { clsx } from 'clsx'
 
@@ -8,9 +9,9 @@ import s from './message-input.module.scss'
 export type MessageInputProps = {
   boxClassName?: string
   buttonProps?: ComponentPropsWithoutRef<'button'>
-} & ComponentProps<'input'>
+} & TextareaAutosizeProps
 
-export const MessageInput = forwardRef<HTMLInputElement, MessageInputProps>(
+export const MessageInput = forwardRef<ElementRef<'textarea'>, MessageInputProps>(
   ({ boxClassName, buttonProps, className, ...restProps }, ref) => {
     const classNames = {
       box: clsx(s.box, boxClassName),
@@ -20,7 +21,7 @@ export const MessageInput = forwardRef<HTMLInputElement, MessageInputProps>(
 
     return (
       <div className={classNames.box}>
-        <input {...restProps} className={classNames.input} ref={ref} />
+        <TextareaAutosize maxRows={10} {...restProps} className={classNames.input} ref={ref} />
         <button aria-label={'Send message'} className={classNames.button} {...buttonProps}>
           <SendMessageIcon />
         </button>