+++ packages/ui-kit/src/components/textarea-autosize/index.ts
@@ -1,0 +1,5 @@
+export {
+  type Style,
+  type TextareaAutosizeProps,
+  default as TextareaAutosize,
+} from './textarea-autosize'