+++ packages/ui-kit/src/hooks/use-fonts-loaded-listener.ts
@@ -1,0 +1,5 @@
+import { useListener } from './use-listener'
+
+export const useFontsLoadedListener = (listener: (event: Event) => any) => {
+  useListener(document.fonts, 'loadingdone', listener)
+}