+++ packages/ui-kit/src/hooks/use-window-resize-listener.ts
@@ -1,0 +1,5 @@
+import { useListener } from './use-listener'
+
+export const useWindowResizeListener = (listener: (event: UIEvent) => any) => {
+  useListener(window, 'resize', listener)
+}