+++ packages/ui-kit/src/utils/index.ts
@@ -1,1 +1,3 @@
 export * from './get-file-size'
+export * from './pick'
+export * from './noop'