+++ packages/ui-kit/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @it-incubator/ui-kit
 
+## 0.2.21
+### Patch Changes
+
+- UIL-82 feat: add autosize textarea to chat input instead of an input
+- 1d4efe6: UIL-82 fix: add input padding when there is selected value
+- 5817a7d: UIL-82 feat: add showUsername option to be able to show who sent the message from current user's perspective
+
 ## 0.2.20
 ### Patch Changes